refactoring(15): decompose MonitorPage.tsx (extract monitorSections)
First slice of req-15 (multi-PR, one component each). MonitorPage.tsx 859 -> 505
lines: extract the module-scope presentational/helper layer (MetricDisplay,
computeBestMetrics, IdeaCardContent — the Experiment-Configuration card) VERBATIM
into components/Monitor/monitorSections.tsx; container imports the two symbols
from it. Moved code byte-identical (modulo export kw); removed the now-dead
ReactNode/AppState/TreeNode/fetchExperimentDetail container imports.

Checks: typecheck 0 non-test errors; build ok; vitest 4 passed / 2 failed
(pre-existing brittle PaperBench tests). req-15 file RETAINED (multi-PR, §9/§10 —
no partial deletion); progress recorded in COMPLETED.md. Remaining: WorkflowPage,
StepResources, SettingsPage, DetailPanel.
