+++ ari-core/ari/viz/frontend/src/README.md
@@ -40,6 +40,7 @@ React/TypeScript source for the ARI `ari.viz` web dashboard — app entry, pages
     - `GpuMonitor.tsx` — GPU usage monitor card.
     - `index.ts` — barrel re-export.
     - `MonitorPage.tsx` — monitor page container.
+    - `monitorSections.tsx` — metric helper + Experiment-Configuration card (extracted from MonitorPage in req 15).
     - `PhaseStepper.tsx` — workflow phase progress bar (idea→bfts→paper→review).
   - `PaperBench/` — register external papers, import them, launch/inspect PaperBench runs.
     - `README.md` — PaperBench index.
